Various bugfixes
- Fixes #178 by setting usesClearTextTraffix = true in config.xml
- Fixes build error when only building for ios and missing
platforms/android dir
- Fixes ios build error caused by missing
NSBluetoothAlwaysUsageDescription (see
https://github.com/dpa99c/cordova-diagnostic-plugin/issues/365)

NOTE: THIS REQUIRES A FULL REINSTALLATION OF THE CORDOVA PLATFORM!
Changes only take effect when /platforms and /plugins are fully cleared
and automatically reinstalled.

+++ hooks/before_compile.js
@@ -24,5 +24,6 @@ module.exports = function (context) {
 
 };
 
-
-fs.copyFileSync("hooks/build-extras.gradle", "platforms/android/build-extras.gradle");
+if (fs.existsSync("platforms/android")) {
+    fs.copyFileSync("hooks/build-extras.gradle", "platforms/android/build-extras.gradle");
+}
